Rename canvasEvents to actions
