+++ src/reducers/gridRef/index.ts
@@ -1,8 +1,8 @@
-import events from "../canvasEvents";
+import ACTIONS from "../actions";
 
 export default function gridRefReducer(state: CanvasAppState, action: CanvasAppAction) {
   switch (action.type) {
-    case events.SET_GRID_REF:
+    case ACTIONS.SET_GRID_REF:
       return action.payload.gridRef;
     default:
       return state.gridRef;
