+++ src/reducers/mode/index.ts
@@ -1,8 +1,8 @@
-import events from "../canvasEvents";
+import ACTIONS from "../actions";
 
 export default function modeReducer(state: CanvasAppState, action: CanvasAppAction) {
   switch (action.type) {
-    case events.CHANGE_MODE:
+    case ACTIONS.CHANGE_MODE:
       return action.payload.mode || state.mode;
     default:
       return state.mode;
