+++ src/reducers/nodesToMove/index.ts
@@ -1,10 +1,10 @@
 import { keys } from "lodash";
 import { isInside } from "../../canvas-utils";
-import events from "../canvasEvents";
+import ACTIONS from "../actions";
 
 export default function nodesToMoveReducer(state: CanvasAppState, action: CanvasAppAction) {
   switch (action.type) {
-    case events.CANVAS_MOUSE_UP: {
+    case ACTIONS.CANVAS_MOUSE_UP: {
       if (state.boxPreview && state.mode === "select") {
         const ids = keys(state.nodes);
         return ids.filter(id => isInside(state.nodes[id], state.boxPreview!.box));
